full profiler
Former-commit-id: 05a1127ffead45cc6ec732789c18a6a9078f9ff7 [formerly e9c93ad8a5ae7bac8f0fc44d94c7f021dbbf7b4b]
Former-commit-id: 8c010e338ab84533baf6c730bd198738d6ccad78

+++ Util/Tokenizer.cpp
@@ -45,7 +45,7 @@ bool Tokenizer::Next(float& token) {
     char* buf = new char[1 + mCurrent - start];
     buf[mCurrent - start] = '\0';
     memcpy(buf, mBuffer + start, mCurrent - start);
-    token = atof(buf);
+    token = (float)atof(buf);
     delete[] buf;
 
     return true;
